+++ app/src/main/java/org/android/go/sopt/ui/signup/SignUpActivity.kt
@@ -1,10 +1,12 @@
-package org.android.go.sopt
+package org.android.go.sopt.ui.signup
 
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
+import org.android.go.sopt.R
 
-class MainActivity : AppCompatActivity() {
+class SignUpActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_sign_up)
     }
 }
