+++ app/src/main/res/layout/activity_sign_in.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:paddingHorizontal="30dp"
-    tools:context=".MainActivity">
+    tools:context=".ui.signin.SignInActivity">
 
 
     <TextView
